refactor(schedule-runs): group components into folders
