+++ src/views/schedule-runs/schedule-runs-table/__tests__/schedule-runs-table.test.tsx
@@ -8,8 +8,8 @@ import {
 import { getMockWorkflowListItem } from '@/route-handlers/list-workflows/__fixtures__/mock-workflow-list-items';
 import { type WorkflowListItem } from '@/route-handlers/list-workflows/list-workflows.types';
 
-import { type Props as RuntimeCellProps } from '../schedule-runs-runtime-cell.types';
-import { type Props as StatusCellProps } from '../schedule-runs-status-cell.types';
+import { type Props as RuntimeCellProps } from '../../schedule-runs-runtime-cell/schedule-runs-runtime-cell.types';
+import { type Props as StatusCellProps } from '../../schedule-runs-status-cell/schedule-runs-status-cell.types';
 import ScheduleRunsTable from '../schedule-runs-table';
 import { type Props } from '../schedule-runs-table.types';
 
@@ -24,12 +24,12 @@ jest.mock('@/components/link/link', () =>
     <a href={href.toString()}>{children}</a>
   ))
 );
-jest.mock('../schedule-runs-runtime-cell', () =>
+jest.mock('../../schedule-runs-runtime-cell/schedule-runs-runtime-cell', () =>
   jest.fn(({ startTime, closeTime }: RuntimeCellProps) => (
     <>{`${startTime} → ${closeTime}`}</>
   ))
 );
-jest.mock('../schedule-runs-status-cell', () =>
+jest.mock('../../schedule-runs-status-cell/schedule-runs-status-cell', () =>
   jest.fn(({ status }: StatusCellProps) => <>{status}</>)
 );
 jest.mock('@/components/table/table', () =>
