+++ src/views/schedule-runs/schedule-runs-table/schedule-runs-table.config.ts
@@ -5,8 +5,8 @@ import { type TableConfig } from '@/components/table/table.types';
 import { type WorkflowListItem } from '@/route-handlers/list-workflows/list-workflows.types';
 import getSearchAttributeValue from '@/views/shared/workflows-list/helpers/get-search-attribute-value';
 
-import ScheduleRunsRuntimeCell from '../schedule-runs-runtime-cell';
-import ScheduleRunsStatusCell from '../schedule-runs-status-cell';
+import ScheduleRunsRuntimeCell from '../schedule-runs-runtime-cell/schedule-runs-runtime-cell';
+import ScheduleRunsStatusCell from '../schedule-runs-status-cell/schedule-runs-status-cell';
 
 export default function getScheduleRunsTableConfig(
   domain: string,
