+++ src/views/schedule-runs/schedule-runs-table/schedule-runs-table.tsx
@@ -2,7 +2,7 @@
 
 import Table from '@/components/table/table';
 
-import getScheduleRunsTableConfig from './config/schedule-runs-table.config';
+import getScheduleRunsTableConfig from './schedule-runs-table.config';
 import { type Props } from './schedule-runs-table.types';
 
 export default function ScheduleRunsTable({
